feat(xml_fixer): Refactor xmlfinder aggregation query
